backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai==1.54.0
httpx==0.27.2
python-dotenv==1.0.0

